+++ docs/implementations.rst
@@ -284,3 +284,17 @@ All network implementations are located in `mockturtle/networks/`:
 +--------------------------------+-------------+-------------+-------------+-------------+-----------------+
 | ``get_output_name``            |             |             |             |             |                 |
 +--------------------------------+-------------+-------------+-------------+-------------+-----------------+
+|                                | *Coloring methods*                                                      |
++--------------------------------+-------------+-------------+-------------+-------------+-----------------+
+| ``new_color``                  |             |             |             |             |                 |
++--------------------------------+-------------+-------------+-------------+-------------+-----------------+
+| ``current_color``              |             |             |             |             |                 |
++--------------------------------+-------------+-------------+-------------+-------------+-----------------+
+| ``color``                      |             |             |             |             |                 |
++--------------------------------+-------------+-------------+-------------+-------------+-----------------+
+| ``paint``                      |             |             |             |             |                 |
++--------------------------------+-------------+-------------+-------------+-------------+-----------------+
+| ``eval_color``                 |             |             |             |             |                 |
++--------------------------------+-------------+-------------+-------------+-------------+-----------------+
+| ``eval_fanins_color``          |             |             |             |             |                 |
++--------------------------------+-------------+-------------+-------------+-------------+-----------------+
