+++ docs/views.rst
@@ -89,3 +89,14 @@ algorithm.  Several views are implemented in mockturtle.
 
 .. doxygenclass:: mockturtle::cnf_view
    :members:
+
+`color_view`: Manages traversal IDs
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+**Header:** ``mockturtle/views/color_view.hpp``
+
+.. doxygenclass:: mockturtle::color_view
+   :members:
+
+ .. doxygenclass:: mockturtle::out_of_place_color_view
+   :members:
