+++ include/mockturtle/views/color_view.hpp
@@ -25,7 +25,7 @@
 
 /*!
   \file color_view.hpp
-  \brief Color view
+  \brief Manager view for traversal IDs, called colors
 
   \author Heinz Riener
 */
@@ -185,6 +185,7 @@ public:
     return values[n];
   }
 
+  template<typename _Ntk = Ntk, typename = std::enable_if_t<!std::is_same_v<typename _Ntk::signal, typename _Ntk::node>>>
   auto color( signal const& n ) const
   {
     return values[this->get_node( n )];
